Drop dependency on GNU readlink
Previously, GNU's `readlink -f` served the purpose of returning the
absolute path to the root source directory of bask, following all
symlinks.

As you can see in the comment in `bin/runner`, [this StackOverflow
answer][1] provides a way of accomplishing this without needing
`readlink -f`. It works for many ways of invoking the script, including

- as a standalone executable (`./bin/runner.sh`)
- being invoked with bash (`bash ./bin/runner.sh`)
- being called from a different directory (`../../bask/bin/runner.sh`)
- when there are symlinks on the path

[1]: http://stackoverflow.com/a/42918