Drop dependency on GNU date
The dependency on GNU date was solely for its ability to print time
stamps with milliseconds. For most things, milliseconds aren't all that
important, though they provide nicer timings than "0 ms" or "1 s".

When we detect that GNU date is available, we use it. Otherwise, we just
use the normal date, and approximate everything to the nearest 1000
milliseconds.