Add semantic aliases for some commands
It makes more sense for a list of dependencies to be called
`bask_depends` rather than `bask_sequence`. The latter can be thought of
as exposing implementation details, while the former is an abstraction.

With `bask_fork_join`, it wasn't readily apparent whether the command
would wait for it's commands to finish before continuing, or whether
they would be spawned in the background.